+++ src/main/java/emanuela/carrubba/viaggi/repositories/ViaggiRepository.java
@@ -3,5 +3,5 @@ package emanuela.carrubba.viaggi.repositories;
 import emanuela.carrubba.viaggi.entities.Viaggi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ViaggiRepository extends JpaRepository<Viaggi, Integer> {
+public interface ViaggiRepository extends JpaRepository<Viaggi, Long> {
 }
